+++ hyundai_kia_connect_api/KiaUvoApiCN.py
@@ -18,7 +18,7 @@ from dateutil import tz
 from .ApiImpl import (
     ClimateRequestOptions,
 )
-from .stampApiImpl import stampApiImpl
+from .ApiImplType1 import ApiImplType1
 
 from .Token import Token
 from .Vehicle import (
@@ -99,7 +99,7 @@ def _check_response_for_errors(response: dict) -> None:
             raise APIError(f"Server returned: '{response['resMsg']}'")
 
 
-class KiaUvoApiCN(stampApiImpl):
+class KiaUvoApiCN(ApiImplType1):
     data_timezone = tz.gettz("Asia/Shanghai")
     temperature_range = [x * 0.5 for x in range(28, 60)]
