--- hyundai_kia_connect_api/stampApiImpl.py
@@ -1,8 +1,0 @@
-from .ApiImpl import (
-    ApiImpl,
-)
-
-
-class stampApiImpl(ApiImpl):
-    def __init__(self) -> None:
-        """Initialize."""
